requirements.txt
django
djangorestframework
requests
xmltodict
PyPDF2
